+++ src/test/java/calc/NumberTest.java
@@ -1,10 +1,9 @@
-package step1;
+package calc;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class NumberTest {
